[feat][2025-09-28]Iteracion 1 y 2 funcionando

+++ src/main/java/com/irene/mercadona/service/AsignacionService.java
@@ -1,37 +1,45 @@
 package com.irene.mercadona.service;
 
+import com.irene.mercadona.dto.AsignacionRequestDTO;
 import com.irene.mercadona.model.Asignacion;
 import com.irene.mercadona.model.Seccion;
 import com.irene.mercadona.model.Trabajador;
 import com.irene.mercadona.repository.AsignacionRepository;
-import com.irene.mercadona.repository.SeccionRepository;
-import com.irene.mercadona.repository.TrabajadorRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+import java.util.Optional;
 
+/**
+ * Lógica de negocio correspondiente a las Asignaciones
+ */
 @Service
 @RequiredArgsConstructor
 public class AsignacionService {
     private final AsignacionRepository asignacionRepository;
-    private final TrabajadorRepository trabajadorRepository;
-    private final SeccionRepository seccionRepository;
+    private final TrabajadorService trabajadorService;
+    private final SeccionService seccionService;
 
     public List<Asignacion> consultarAsignaciones() {
         return asignacionRepository.findAll();
     }
 
-    public Asignacion consultarPorCodigo(Long codigo) {
-        return asignacionRepository.findById(codigo)
-                .orElseThrow(() -> new RuntimeException("Asignación no encontrada"));
+    public Optional<Asignacion> consultarPorCodigo(Long codigo) {
+        return asignacionRepository.findById(codigo);
     }
 
-    public Asignacion crearAsignacion(String dniTrabajador, Long codSeccion, int horas) {
-        Trabajador trabajador = trabajadorRepository.findById(dniTrabajador)
-                .orElseThrow(() -> new RuntimeException("Trabajador no encontrado"));
-        Seccion seccion = seccionRepository.findById(codSeccion)
-                .orElseThrow(() -> new RuntimeException("Sección no encontrada"));
+    public Asignacion crearAsignacion(AsignacionRequestDTO asignacionRequestDTO) {
+        Trabajador trabajador = trabajadorService.consultarTrabajadorPorDni(asignacionRequestDTO.getDniTrabajador())
+                .orElseThrow(() -> new IllegalArgumentException("{error.trabajador.no_encontrado}"));
+        Seccion seccion = seccionService.consultarSeccionPorCod(asignacionRequestDTO.getCodSeccion())
+                .orElseThrow(() -> new IllegalArgumentException("{error.seccion.no_encontrada}"));
+
+        int horas = asignacionRequestDTO.getHorasAsignadas();
+
+        int horasAsignadas = asignacionRepository.sumHorasByTrabajadorId(asignacionRequestDTO.getDniTrabajador()).orElse(0);
+        if (horasAsignadas + horas > trabajador.getHorasDisponibles()) {
+            throw new IllegalArgumentException("{error.trabajador.horasmax}");
+        }
 
         Asignacion asignacion = new Asignacion();
         asignacion.setTrabajador(trabajador);
@@ -41,27 +49,12 @@ public class AsignacionService {
         return asignacionRepository.save(asignacion);
     }
 
-    public Asignacion editarAsignacion(Long codigo, String nuevoTrabajadorDni, Long nuevaSeccionCod, int nuevasHoras) {
-        Asignacion asignacion = consultarPorCodigo(codigo);
+    public boolean eliminarAsignacion(Long cod) {
 
-        if (nuevoTrabajadorDni != null) {
-            Trabajador trabajador = trabajadorRepository.findById(nuevoTrabajadorDni)
-                    .orElseThrow(() -> new RuntimeException("Trabajador no encontrado"));
-            asignacion.setTrabajador(trabajador);
+        if(asignacionRepository.existsById(cod)) {
+            asignacionRepository.deleteById(cod);
+            return true;
         }
-
-        if (nuevaSeccionCod != null) {
-            Seccion seccion = seccionRepository.findById(nuevaSeccionCod)
-                    .orElseThrow(() -> new RuntimeException("Sección no encontrada"));
-            asignacion.setSeccion(seccion);
-        }
-
-        asignacion.setHorasAsignadas(nuevasHoras);
-
-        return asignacionRepository.save(asignacion);
-    }
-
-    public void eliminarAsignacion(Long cod) {
-        asignacionRepository.deleteById(cod);
+        return false;
     }
 }
